ProductDisplayScreen file renamed to ProductsBrowseScreen
