+++ frontend/src/components/Team.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Jumbotron, Row, Col} from 'reactstrap';
 import Zoom from 'react-reveal/Zoom'
-import '../../style/2/Team.css'
+import '../style/2/Team.css'
 
 const Team = () => {
     return (
